requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
aiomysql==0.2.0
telethon==1.37.0
cryptography==46.0.5
